pyproject.toml
[project]
name = "free-ais"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.6"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "load-dotenv>=0.1.0",
    "requests>=2.32.3",
    "websockets>=14.1",
]

[dependency-groups]
dev = [
    "ruff>=0.8.1",
]
